backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
aiohttp>=3.9.0
websockets>=12.0
adbutils>=1.2.0
python-multipart>=0.0.6
scrapy>=2.11.0
scrapy-playwright>=0.0.34
scrapy-user-agents>=0.1.1
scrapy-rotating-proxies>=0.6.2
